sbn-django/requirement.txt
Pillow==8.3.2
djangorestframework==3.12.4  
django-cors-headers==3.8.0
django-phonenumber-field[phonenumbers]==5.2.1.dev7+g2fa60ad
djangorestframework-simplejwt==4.8.0.post2+g0e98508
pyjwt==2.1.0